refactor(dto): extract workspace IPC contracts
- move Workspace request and response types into the DTO layer

- keep command handlers focused on transport delegation
